gtk-demo: Avoid crash in CSS demos
Avoid a potential crash in the CSS demos where the text editing buffer is
out of sync with the last good CSS parsing buffer. In the case of CSS
warnings, we get a parsing-error callback but no exception is raised.
This would cause the buffers to become out of sync and accessing position
information from the parsing-error section would crash the text editor due
to an out of range iterator being created.