Ignore GValueArray deprecations
Wrap calls to GValueArray related calls with G_GNUC_BEGIN/END_IGNORE_DEPRECATIONS.
Although GValueArray is deprecated, we still need to support the marshaling of
them in PyGObject. The deprecations add noise to the build processes in which
new warnings could be lost. Essentially losing the element of surprise a new
warning should have on maintainers.