Add marshalling coercion for Python classes and instances to GTypeClass
Automatically marshal Python GObject classes and instances to GTypeClass
structs (GObjectClass). This allows usage of the GTypeClass methods by
passing a Python GObject class or instance to the GTypeClass method.
This is needed to support usage of GTypeClass methods since we don't
manually bind GTypeClasses and they are not very well supported with
introspection.

https://bugzilla.gnome.org/show_bug.cgi?id=685218