Split the callable cache into the different types
Instead of doing different things based on the various function types
this adds vfuncs for generate_args_cache() and invoke() which are then
specialized for the various function types. Also add a calling context
to the callable cache which is then used to determine the direction
when generating the arg caches.

This is in preparation for adding closure caches.

https://bugzilla.gnome.org/show_bug.cgi?id=727004