Use the caches for closures, but not yet for marshaling the arguments
Instead of using the various GI functions we use the data from the caches.
This also fixes generating an arg cache for a closure as it was missing some
data or simply setting incorrect data. Also, always included the GITypeInfo
until the closures no longer need it for marshaling the arguments.

https://bugzilla.gnome.org/show_bug.cgi?id=727004