Use the caches for marshaling the arguments in closures
Changes to object marshaling when the calling context is from
C was required to correctly keep the correct floating status.

The array cache has been modified to set to/from_py_marshaller
for the length arg cache. This is required for closures which include
the length arg for backwards compatibility. The closure cache takes
care to change the length arg cache's meta type so it gets marshalled
for closures.

https://bugzilla.gnome.org/show_bug.cgi?id=727004