Refactor signal array length marshalling to support a length policy
Replace passing arrays of data to non-caching array marshallers with a policy
closure that can be customized depending on context. In the case of signals,
this is mostly scaffolding which will be replaced with caching marshallers.
However, it opens the legacy marshaller for usage with struct and object
array fields.

https://bugzilla.gnome.org/show_bug.cgi?id=688792