Explicitly check if an override exists instead of ImportError
If an override depended on another module and it did not
exist then the raised ImportError was consumed and assumed
to mean that the override did not exist. This makes it
difficult to diagnose issues with overrides.

This uses pkgutil.get_loader() as it is the easier way to
determine if a module exists in both Python 2 and 3
and avoid deprecated functions.

https://bugzilla.gnome.org/show_bug.cgi?id=749532