tests: Update Makefile for building tests on OS X
Pull LDFLAGS into variables for command and Python extension specific.
Use -shext .so for the Python extension on both Linux and OS X. The .so is
is needed by Python on OS X (not dylib).
Remove "all" target and explicit .la targets as these are handled by automake
and the "check" target on demand.

https://bugzilla.gnome.org/show_bug.cgi?id=762176