build: Do not enable code coverage based on lcov
Code coverage turns on various slow paths and disables optimizations. It
should be, and it usually is, an opt in configuration flag.

Enabling it should not be conditional on the existence of the lcov
binary on the system, otherwise anybody building PyGObject from Git on
any moderately complete development environment will automatically
enable code coverage even when not wanted.

https://bugzilla.gnome.org/show_bug.cgi?id=764075