Move pep8/pyflakes tests from 'make check' to 'make check.quality'
The current behaviour is problematic because new versions of pep8/pyflakes
add new warnings or get more picky. This makes testing
hard in case newly added code is developed on an older version
or when testing old branches/releases.

This moves the pep8/pyflakes tests to a new "make check.quality"
command.

This also removes the SKIP_PEP8 flag which was added in 25d12afd06863ce
to speed up make check.

https://bugzilla.gnome.org/show_bug.cgi?id=764087