Remove pygobject-private.h and rename pygobject.c to pygobject-object.c
Move all the random declarations in pygobject-private.h to their
respective header files. Rename pygobject.c to pygobject-object.c
so it's clearer that it's not the implementation of pygobject.h.
Add a new pygobject-internal.h which includes pygobject.h
with _INSIDE_PYGOBJECT_ defined like pygobject-private.h did.

In case you are looking at the git log and end up here due to the
rename try:

  git log --follow pygobject-object.c

or on the web interface go to the history of the old file name:

https://git.gnome.org/browse/pygobject/log/gi/pygobject.c?id=6b702c052e9f26e809cff494f0c896d17a514c64

https://bugzilla.gnome.org/show_bug.cgi?id=767084