Add support for bytes and non-utf-8 file names.
Py2+Unix: Convert unicode to bytes using the fsencoding. Pass bytes as is.
Returns path as bytes as is.

Py2+Windows: Convert unicode to wtf-8. Pass bytes as is.
Returns path as bytes as is.

Py3+Unix: Convert str using os.fsencode so that the surrogateescape handler
can restore the real path if the source was a Python API such as os.listdir
sys.argv etc. Pass bytes as is. Return str decoded using os.fsdecode so that
it can be passed to Python API such as open, os.listdir etc.

Py3+Windows: Convert str to wtf-8. Decode bytes using the fsencoding first.
Returns str + surrogates.

This change makes anyting taking filenames on Python 3 behave the same
as Python functions like listdir() or open(). Compared to Python 3 builtin
functions which return the same type that was passed in we always return str.

One remaining problem is that glib assumes that Windows paths are utf-16,
while they are just 16bit arrays and as a result utf-8 is not enough to
represent all possible paths. We use wtf-8 here instead (utf-8 with lone surrogates),
which allows us to convert all paths, but glib functions accessing the fs
will fail with them.

PyUnicode_EncodeFSDefault was added in CPython 3.2 so bump the requirement.

https://bugzilla.gnome.org/show_bug.cgi?id=746564