Use optimized version of PyImport_ImportModule
https://gitlab.gnome.org/GNOME/pygobject/issues/123

You would expect PyImport_ImportModule to import a module once, store
it's reference to sys.modules and the next time we call it with the same
argument reuse the reference from sys.module. But it is not what is
happening unfortunately.

Everytime we call it it releases GIL, grabs global python import lock,
grabs GIL back. The proposed optimization looks at sys.modules first and
uses the reference from there to avoid GIL congestion. pyg_type_lookup is
a good example of a function which imports python modules at run time and
can cause GIL congestion.