gi/pygtype.c: avoid GIL congestion in pyg_type_lookup
https://gitlab.gnome.org/GNOME/pygobject/issues/123

99% of boxed gtypes will not be registered through pyg_register_gtype_custom
API. GStreamer API is very good example. So pyg_type_lookup will return NULL
most of the times. The optimization takes advantage of that improving
performance significantly. The most important improvement is - we are not
calling pygi_type_import_by_g_type in the loop, which can lead to GIL
congestion