aclocal: make local file discover by reading AC_CONFIG_MACRO_DIR work
aclocal needs to know where the local macros are for including them and
for having a place to put the required system ones in.

Since 1.13 it can parse configure.ac and get the path from AC_CONFIG_MACRO_DIR.
This didn't work because for some reasons it complains if the local macros
start with "AM_" even if they are there (I guess it assumes they can't be local).

To work around that, change the AM prefix to PYG, bump the automake version so
we can asume aclocal finds it and remove the hardcoded paths from both
autogen.sh and Makefile.am.

https://bugzilla.gnome.org/show_bug.cgi?id=777713