tests: Remove TestMainLoop.test_concurrency
The code containing the bug no longer exists and the test doesn't do
what it says. The thread never gets terminated and it doesn't
assert anything. And I'm not sure what it should assert.

For the corresponding bug see
    https://bugzilla.gnome.org/show_bug.cgi?id=663068

While the test doesn't hurt, it adds half a second wait time
to the test suite.

https://bugzilla.gnome.org/show_bug.cgi?id=698548