Fix conversion from pointers to hashfunc return values.
Fixes comiler warnings on 64bit Windows.

Python 3 now uses Py_hash_t as return value for hashfunc. Use
a compat define so we can use the right type depending on the
Python version.

On Python 2 + 64bit Windows the pointer doesn't fit the long
return type. Add a macro which handles the casting for that case
to prevent compiler warnings.

https://bugzilla.gnome.org/show_bug.cgi?id=780591