tests: Fix make check.valgrind
Got broken when we changed to use dbus-run-session. Move
valgrind related env vars before the dbus launcher helper.

This also tries to improve the lookup of the suppression file.
Instead of relying on the basename of the python binary use the python
major version to select one of two suppression files in the repo.

Add PYTHONMALLOC=malloc when running under valgrind, which makes Python
use malloc instead of its own allocator. Only works with Python 3.6+.