gi/pygi-value: Don't wrap GValue in GValue when creating GValueArray
If a GValueArray is created from a list of GValues, it should not wrap
these GValues in a second layer of GValues before appending to the
array. The end result should be a GValueArray of GValues, not a
GValueArray of GValues holding another GValue (unless the user
explicitly creates such a value).

With this patch the behavior is now consistent when creating a
GValueArray and appending GValues directly, and creating a GValueArray
within a GValue based on a list of GValues.

For instance, to create a GValueArray of G_TYPE_UINT the user must
create GValues manually and create a GValueArray from these. The result
should be a GValueArray of GValues that hold G_TYPE_UINT, not a
GValueArray that contain GValues that hold a GValue that holds
G_TYPE_UINT.

See !66