Fix potential uninitialized memory access during GC
We use _PyGIDefaultArgPlaceholder as a sentinel value to represent default
values during function argument list construction. Right now, it's a Python
type object. We make it using PyObject_New, so most of its fields end up
uninitialized. The object body being uninitialized wouldn't be a problem if
the placeholder object were unreachable, but the object *can* be reached
during GC by traversal through frame objects.

Depending on the exact contents of the uninitialized memory, the GC can go on
to cause other kinds of memory corruption through the process.

IMHO, the easiest fix for this problem is to just make the placeholder a
simpler data structure, like a list.

https://bugzilla.gnome.org/show_bug.cgi?id=786872