setup.py: Port to distutils/setuptools
Instead of wrapping autotools add a proper setuptools based build system.
Compared to the autotools one this does not install .pc files or headers
and does not allow running tests.

It uses pkg-config for discovering dependencies and explictely searches
for .pc files in the Python prefix so that pycairo installations in a
virtualenv are discovered. When using MSVC, pkg-config is skipped and
it is assumend that INCLUDE and LIB is properly set up.

Version information and requirements are parsed from configure.ac, package
metadata is parsed from PKG-INFO.in.

Also adds a "setup.py distcheck" command which makes sure all tracked files
end up in the tarball and that the tarball builds (no tests are run atm).

https://bugzilla.gnome.org/show_bug.cgi?id=789211