PKG-INFO: Revert name back to PyGObject
My thinking there was that it gives us the same naming of tarballs
as with the autotools build system. But on a second thought, we
can't use the same tarball anyway due to gz vs xz and while the
naming should be case insensitive it's not worth the risk.

The commit changing it was 3e455944f5835c7509