Add sphinx based documentation
Takes the documentation developed at
https://github.com/pygobject/pygobject-docs
and puts into /docs

See https://pygobject.readthedocs.io for how it looks

Now that we move to gitlab we can use webhooks to trigger builds
on readthedocs from gitlab directly and we should also have a nicer
contribution UX.

This also gets rid of most of README/HACKING/INSTALL and moves most
of the information into the documentation. The README is kept short
and only makes clear what pygobject is and points to the online docs
as that should answer all questions.

setup.py now sets the content of README.rst as long_descriptions,
as that is the content shown on PyPI. This makes the page on PyPI
look the same as on gitlab.

https://bugzilla.gnome.org/show_bug.cgi?id=791448