docs: convert NEWS file to reST and link bugs
Looks nicer in the online docs and makes it easier to look
up bugs and link to specific releases.

Rename the custom reST role for bugs from gnomebug to bzbug.
With the upcoming move to gitlab that makes it more clear that those
are referencing bugzilla.

Adjust the "make release-news" target to output something similar
to the new format to make converting to reST easier.

Update the maintainer release todo list to state that the changelog
should be copied from stable branches to master. It would be
nice if the NEWS file in master, which gets exposed on readthedocs,
would contain all the information available.