docs: Update for the git repo move from git.gnome.org to gitlab.gnome.org
Changes all the links to point to the gitlab instance.

Adds the gitlab setup to the sphinx theme configuration, so the theme
can show "Edit on GitLab" links on each page.

Create a new page pointing to the git report and bug tracker.
Someone on IRC mentioned that this info is hard to find, so try to make
it more proiminent.