overrides: Fix Gtk.Adjustment.__init__ overrides not setting "value" sometimes. Fixes #151
Gtk.Adjustment allows passing positional arguments to __init__ which get
translated to a dict for passing to GObject.Object.__init__. In case of the
first argument "value", if "value" is passed before the upper and
lower bound to Object.__init__ it will be set to 0 instead.

In Python 2 this happened to work (at least on my machine) because
"value" got placed after the bounds (in terms of iteration order)
in the final dict value passed to Object.__init__.

To work around this, set "value" again after __init__().
A similar work around already exists when "value" is passed as a kwarg.