autotools: use flake8 instead of pep8/pyflakes
flake8 uses pep8/pycodestyle/pyflakes internally, so there should not
by any difference.

The nice thing about it is that we can store the configuration in
setup.cfg and linting comes down to just executing "flake8".
It also allows to ignore warnings with inline comments, so that
when avoiding a warning isn't possible it doesn't need to be disabled
for the whole codebase.