setup.py: Make the test command work under MSYS2
gobject-introspection requires an import lib (.dll.a) on Windows,
so build that as well.

Disable the dbus tests until we have a temp bus like with autotools.
Unset MSYSTEM to disable the os.sep hackery.

This hit a bug in GI for the mingw+no-libtool case, see
https://bugzilla.gnome.org/show_bug.cgi?id=791902
The fix has been applied to MSYS2 in the meantime, see
https://github.com/Alexpux/MINGW-packages/pull/3308