setup.py: Add logging for pycairo lookup and fail early if the header isn't found. See #150
Log the pycairo module we find and the include path we deduce from that.
Also check that the directory actually includes the right header file
instead of letting the compiler fail later on.

This will hopefully make it more clear why installing through pip isn't working
in some cases (see #150 for some reports)