setup.py: rework pycairo header lookup and use new pycairo API when available. See #150
pycairo git/next version now has a get_include() function which gives us the
include path matching the Python module in all cases.

Rework the lookup code to try that API if possible and fall back to the old
way. Also add logging everywhere so it's easier to debug if there is still
something wrong.