gitlab-ci: use pyenv in docker instead of the deadsnakes PPA
deadsnakes only provides packages for Ubuntu LTS versions and doesn't
include versions present in that Ubuntu version.
With Ubuntu zesty this happened to work, but now that that's EOL
we have to switch to artful and no longer have a Python 3.5 available.

Instead switch to using pyenv in docker and compiler our own Python
versions. This should make it easier to swtich distros in the future.

Also adds a run-docker.sh script which builds the image and
runs it with the git repo mounted. This should make local testing
easier in the future.