tests: remove locale test for uppercasing C names
Someone on IRC had problems running the test suite, and that's not the first
time, so remove it for now (added in 63c7f17c2248)

The locale generation adds extra dependencies and only works on Linux,
and while the test does something useful it complicates things too much.

We need to find a better way (in Python if possible) to test for locale
dependend upper()/lower().