tests: add a pytest hook for handling unhandled exception in closures
In PyGObject when an exception is raised in a closure called from C then
the error gets passed to sys.excepthook (on the main thread at least)
and the error is by default printed to stdout. Since pytest by default
hides stdout, errors can be easily missed now.

To make these errors more visible add a test wrapper which checks
sys.excepthook for unhandled exceptions and reraises them.
This makes the tests fail and as a bonus also shows the right
stack trace instead of just the error message.