Store the closure cache in the arg cache. Fixes #103
Instead of creating the closure cache on every call store
it in the arg cache.

The only potential problem here is that the arg cache owns the
closure cache and needs to stay around. But afaics the cache
stays around until interpreter shutdown and the closure
code guards against that and will simply do nothing then.
This would all be easier if the caches were refcounted.