struct: fetch the base info from the struct type and not the instance
We need the base info in the struct tp_dealloc and PyPy doesn't like
when the instance escapes there by passing it to some Python API.

Since the base info is stored on the class anyway get it from there instead.

This allows the test suite to run without crashes under PyPy.