PyGObject: rework toggle ref logic to not stricly require tp_dictoffset so it works under PyPy
For toggle refs we create a custom instance dict on demand and when that gets triggered
enable toggle refs. Under PyPy this doesn't work as it doesn't fully support creating a
custom instance dict and tp_dictoffset. This leads to wrappers getting wrongfully GCed
and their state getting lost.

To work around this, instead of checking if a custom dict was created, check if the default
dict is not empty instead (PyPy only), and enable toggle refs then.