Add a minimal implementation of Gtk.Template. See #52
This tries to add a minimal API which allows for basic template usage
with the possibility to maybe add more automation/subclassing/nesting
later on.

Compared to gi_composites.py this adds parameters to Child and Callback
to set the name, init_template() doesn't need to be called and is stricter
in what it supports to allow future improvements.

The _gtktemplate.py file should be resuable with older PyGObject versions
with the only difference that init_template() needs to be called.