setup.py: Make lookup of Python valgrind suppression files work on Fedora and in a venv
They are in /usr/share/doc/python(2-devel|3-devel|\d\d) on Fedora.
Also search for the suppression files in more prefixes in case we are in a virtualenv
or in case the Python interpreter isn't installed.

Also removes the .supp files included in the test suite as they are no longer used.