int converters: unify overflow error handling
In case functions PyLong_AsLong() fails we bail out and just return the
error set by it. In most cases this is due to overflow and results in error
messages such as "too large to convert" and "too large to convert to C long".

The problem here is that for the same type of error we get different error
messages depending on the Python version, the architecure (long size) and the
value itself.

To unify the error message reraise the OverflowError with our default
error message in all cases.