Destroy custom GLib.Source instances when they get freed. See #193
When we create a custom GLib.Source instance and attach it to a MainContext
the methods will be called by the context. In case the Python wrapper
gets freed this means the object used in the callbacks is no longer valid
and things crash.

To prevent this case simply destroy the source in case it gets deallocated
so that its methods will no longer be called after that. This means the
user has to make sure that the object stays alive, like now, but without
having crashes if one doesn't.