importer: raise ImportError in load_module() and not find_module(). See #213
find_module() should either return None or a loader, but we raised ImportError
there in case we already knew that the namespace was missing.

Move that check to load_module() instead. While there shouldn't be any functional
difference, raising in find_module() under Python 3 resulted in a chained
exception with an unrelated error message printed first.