gi/pygi-argument: Squash critical warning when using array as hash value
Using a GPtrArray as the value type of a GHashTable currently produces a
warning:
   Unsupported type array
for example when calling a function whose return type is annotated as:
   (transfer container) (element-type filename GPtrArray<SomeObject>)

Since we don’t treat objects (GI_TYPE_TAG_INTERFACE) specially when
marshalling to/from GHashTable, we shouldn’t need to treat arrays
specially either. Handle the case explicitly.

Signed-off-by: Philip Withnall <withnall@endlessm.com>