ci: build python 2.7 and 3.7 in debug mode. Fixes #243
It contains more strict checks which might be useful for finding potential bugs.

Initialize PyGILState_STATE because gcc can't figure out that state is always
defined in the release case when Py_DEBUG is enabled..

Remove -Winline which is triggerd for pygobject_init() with the debug build.
Not sure what to do when inlining fails, so just remove the warning for now.