tests: don't assume anything about sys.excepthook
On Ubuntu (at least the 18.10 dev version) sys.excepthook is some apport_excepthook
object and not the default Python one. Our pytest extension was checking
that sys.excepthook is the default which made every test fail as a result.

Be less strict and don't assume anything about sys.excepthook instead.