meson: Don't link against libpython on non-Windows systems. Fixes #253
While on Windows it needs to link against it, on Unix systems
it doesn't have to if the process loading it has all the symbols
already loaded (python executable itself, or something dlopening
libpython with RTLD_GLOBAL before).

While on Fedora things get linked by default that's not the case with
Debian based distros and macOS, so most software has to deal with it
anyway and we ignore Fedora here for now.

Ideally this should get handled in meson itself, see:
https://github.com/mesonbuild/meson/issues/4117