meson builds: Support Visual Studio as well
Check for the Cairo headers and libraries on Visual Studio builds if the
pkg-config files could not be found, as we don't have direct control
over the build system for Cairo/Cairo-GObject for Visual Studio.

Also, don't add gcc/CLang-specific compiler flags for Visual Studio
builds, but just force-include msvc_recommended_pragmas.h, which is
shipped with GLib, which should make the compiler-time checks that we
really want to look out for.