distutils: Fix and improve Visual Studio builds
The compiler instance for MSVC do not have a compiler sub-attribute, so
we only use that attribute when we are not building with MSVC.  Also,
force-include msvc_recommended_pragmas.h (which one gets from GLib) so
that we also check for warnings we want to look out for like we do for
the GCC/CLang builds (and silence others that don't matter much).

Also enable tests to build and run on Visual Studio builds as well.