setup.py: Fix MSVC builds
The distutils C Compiler object for MSVC does not have a compiler
sub-attribute, so check the compiler type before we try to use that
sub-attribute.  This will fix builds on MSVC using distutils.

Also, in place of checking the various GCC/CLang compiler flags for
compile-type warnings we want to look out for, force-include
msvc_recommended_pragmas.h when we are building with Visual Studio, as
we are doing in the other GNOME projects, to achive similar effects.