overrides: Fix length parameter when passing non-ascii strings to Gtk.Builder.add_from_string/add_objects_from_string
It was passing the codepoint count instead of the length in bytes which lead to things
reading past the provided memory and crashes. Also handle included NULL bytes by only
passing the length up to the NULL byte (also leads to crashes as it's unexpected).

Also use the same code for Gtk.UIManager.add_ui_from_string() which was already passing the
right count but didn't handle embedded NULL bytes.

Fixes #255