GValue: fall back to the custom C marshaller to support fundamental types. Fixes #280
In gst-python a custom gvalue converter is registered for GstFraction which is a fundamental type.
In case of Value.set_value/get_value we only support a limited set of types and only fall back
to the C marshaller which looks up the registered converters in case of boxed types.

Instead of raising for an unhandled type fall back to the custom C marshaller.