gtk: Fix rows getting inserted on the wrong level if parent=None. Fixes #281
TreeStore.insert_before allows passing parent as None in which case the parent
is derived from the sibling. The new code using insert_with_values() didn't take
this into account and simply passed the None along resulting in the row always getting
added at the toplevel.

Fix this by always passing a parent iter to insert_with_values() except when sibling is also
None and we want to add to the toplevel anyway.