tests: add a test for stripping return values from event methods. Fixes #85
As pointed out in #85 we don't use strip_boolean_result() for the main
Gdk.Event struct while we do for the member structs.

But at this point int time I'm afraid that changing it will break existing code,
so just add some tests to document the difference and make sure it doesn't change.