gtk: raise in case Gtk.Window is instantiated after init failed. See #298
gtk crashes if gtk_init() fails and we can't do much about it.
This assumes the Gtk.Window() is the first instantiated widget and
raises an exception in case init failed.

We already had such a check but it was removed in 86a37d67455d
for some reason (maybe because it only helps in some cases)