docs: getting started: default to mingw64 for the MSYS2 instructions
The idea was to default to mingw32 because mingw64 builds had more bugs initially
and if one wants to bundle things mingw32 makes it run on more machines.

But, the gtk stack should be in good shape now with mingw64 and people tend to
prefer 64bit builds (see #326 and #321) ignoring the docs, so lets just default
to mingw64 instead.