gtk overrides: Fix template hierarchy issue
When a widget is inside a template it is created through a
g_object_new and does not have a python wrapper when
pygobject__g_instance_init is called. In that case, a wrapper is
created and the "__init__" method is called to instantiate it. Then,
"init_template" is called to init its own template (if it exists).

However, "init_template" needs to be called before the object
constructor in order to create and instantiate all its children,
signals and properties.

This issue is fixed by calling init_template before the contructor if
the python object has been created through g_object_new.

A new test for the template hierarchy is added (based on an example
from Marinus Schraal).

Closes: #257, #386