gtktemplate: Do not crash on multiple init_template calls
init_template method is automatically called when a widget is
created and it is not supposed to be called multiple times. That is
why this method is turned into a no-op after its first call via a
lambda function.

init_template can still be called after a widget creation (and it is
supposed to do nothing). However, a second call will result in a crash
because the lambda function has a parameter while init_template is not
supposed to have any parameter.

This issue is fixed by removing the parameter of the lambda function.

A new test is also added to assert that a second init_template call
does nothing.