basictype: Handle unicode strings when converting to g(u)int8
This wasn't explicitly handled before and passing unicode would
fail because PyBytes_Check would return 0.

Apparently we never hit this code path, even though setting
unicode properties on things that expected a single byte
was actually tested.

Anyway, we now try to handle this in a "nicer" way - if
the unicode string can be converted into a string of bytes
as UTF-8 and that only takes up one byte, then we can set
the property.